feat : 장비 툴팁
